fix: add scroll event to header
