+++ packages/frontend/src/app/components/header.tsx
@@ -2,6 +2,7 @@
 import { useState } from 'react'
 import Link from 'next/link'
 import { Navigation } from '@/app/components/navigation'
+import { useScrollDirection, ScrollDirection } from '@/app/utils/custom-hook'
 import { CrossIcon, HamburgerIcon, SearchIcon } from '@/app/icons'
 import { SUBSCRIBE_URL, SEARCH_PLACEHOLDER } from '@/app/constants'
 import styles from './header.module.css'
@@ -47,6 +48,7 @@ const AboutUsBtn = (
 export const StickyHeader = () => {
   const [isHamburgerClicked, setIsHamburgerClicked] = useState(false)
   const [isSearchClicked, setIsSearchClicked] = useState(false)
+  const scrollDirection = useScrollDirection()
 
   const onHamburgerOverlayOpen = () => {
     setIsHamburgerClicked(true)
@@ -189,7 +191,10 @@ export const StickyHeader = () => {
 
   return (
     <div
-      style={{ zIndex: '999' }}
+      style={{
+        zIndex: '999',
+        display: scrollDirection === ScrollDirection.UP ? 'flex' : 'none',
+      }}
       className="w-screen flex justify-between fixed top-0 bg-white"
       id="sticky-header"
     >
